Break out data structures to separate files
Standard and unity builds
